+++ ignite/pkg/cosmosutil/genesis/genesis.go
@@ -2,6 +2,7 @@ package genesis
 
 import (
 	"context"
+	"fmt"
 	"os"
 
 	"github.com/ignite/cli/ignite/pkg/jsonfile"
@@ -24,6 +25,8 @@ const (
 	FieldConsensusRootHash           = "app_state.monitoringp.params.consumerConsensusState.root.hash"
 	FieldConsumerUnbondingPeriod     = "app_state.monitoringp.params.consumerUnbondingPeriod"
 	FieldConsumerRevisionHeight      = "app_state.monitoringp.params.consumerRevisionHeight"
+
+	fieldModuleParamFormatString = "app_state.%s.params.%s"
 )
 
 type (
@@ -37,7 +40,12 @@ type (
 	gentxs []struct{}
 )
 
-// FromPath parse genesis object from path.
+// ModuleParamField returns the field name of a given module param pair.
+func ModuleParamField(module, param string) string {
+	return fmt.Sprintf(fieldModuleParamFormatString, module, param)
+}
+
+// FromPath parses genesis object from path.
 func FromPath(path string) (*Genesis, error) {
 	file, err := jsonfile.FromPath(path)
 	return &Genesis{