+++ ignite/services/chain/chain.go
@@ -10,6 +10,7 @@ import (
 	"github.com/tendermint/spn/pkg/chainid"
 
 	chainconfig "github.com/ignite/cli/ignite/config/chain"
+	chainconfigv1 "github.com/ignite/cli/ignite/config/chain/v1"
 	"github.com/ignite/cli/ignite/pkg/chaincmd"
 	chaincmdrunner "github.com/ignite/cli/ignite/pkg/chaincmd/runner"
 	"github.com/ignite/cli/ignite/pkg/cliui/colors"
@@ -307,9 +308,11 @@ func (c *Chain) DefaultHome() (string, error) {
 	if err != nil {
 		return "", err
 	}
-	validator := config.Validators[0]
-	if validator.Home != "" {
-		return validator.Home, nil
+	if len(config.Validators) > 0 {
+		validator := config.Validators[0]
+		if validator.Home != "" {
+			return validator.Home, nil
+		}
 	}
 
 	return c.appHome(), nil
@@ -382,11 +385,14 @@ func (c *Chain) KeyringBackend() (chaincmd.KeyringBackend, error) {
 	}
 
 	// 2nd.
-	validator := config.Validators[0]
-	if validator.Client != nil {
-		if v, ok := validator.Client["keyring-backend"]; ok {
-			if backend, ok := v.(string); ok {
-				return chaincmd.KeyringBackendFromString(backend)
+	if len(config.Validators) > 0 {
+		validator := config.Validators[0]
+
+		if validator.Client != nil {
+			if backend, ok := validator.Client["keyring-backend"]; ok {
+				if backendStr, ok := backend.(string); ok {
+					return chaincmd.KeyringBackendFromString(backendStr)
+				}
 			}
 		}
 	}
@@ -443,10 +449,13 @@ func (c *Chain) Commands(ctx context.Context) (chaincmdrunner.Runner, error) {
 		return chaincmdrunner.Runner{}, err
 	}
 
-	validator := config.Validators[0]
-	servers, err := validator.GetServers()
-	if err != nil {
-		return chaincmdrunner.Runner{}, err
+	servers := chainconfigv1.DefaultServers()
+	if len(config.Validators) > 0 {
+		validator := config.Validators[0]
+		servers, err = validator.GetServers()
+		if err != nil {
+			return chaincmdrunner.Runner{}, err
+		}
 	}
 
 	nodeAddr, err := xurl.TCP(servers.RPC.Address)