+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### Features
 
+- [#3142](https://github.com/ignite/cli/pull/3142) Add `ignite network request param-change` command.
 - [#3181](https://github.com/ignite/cli/pull/3181) Addition of `add` `remove` commands for `plugins`
 - [#3184](https://github.com/ignite/cli/pull/3184) Separate `plugins.yml` config file.
 - [#3038](https://github.com/ignite/cli/pull/3038) Addition of Plugin Hooks in Plugin System
@@ -16,6 +17,7 @@
 - [#2995](https://github.com/ignite/cli/pull/2995/) Add `ignite network request remove-validator` command.
 - [#2999](https://github.com/ignite/cli/pull/2999/) Add `ignite network request remove-account` command.
 - [#2458](https://github.com/ignite/cli/issues/2458) New `chain serve` command UI.
+- [#2992](https://github.com/ignite/cli/issues/2992) Add `ignite chain debug` command.
 
 ### Changes
 
@@ -40,6 +42,7 @@
 - [#3084](https://github.com/ignite/cli/pull/3084) Add Ignite Chain documentation.
 - [#3109](https://github.com/ignite/cli/pull/3109) Refactor scaffolding for proto files to not rely on placeholders.
 - [#3106](https://github.com/ignite/cli/pull/3106) Add zoom image plugin.
+- [#3183](https://github.com/ignite/cli/pull/3183/) Make config optional for init phase.
 - [#3224](https://github.com/ignite/cli/pull/3224) Remove grpc_* prefix from query files in scaffolded chains
 - [#3229](https://github.com/ignite/cli/pull/3229) Rename `campaign` to `project` in ignite network set of commands
 - [#3244](https://github.com/ignite/cli/pull/3244) updated actions.yml for resolving deprecation message